Moved /users/** endpoints to authorized controller
