+++ src/main/java/com/infoshareacademy/responsibledrinkersweb/config/WebSecurityConfig.java
@@ -47,10 +47,9 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     protected void configure(HttpSecurity http) throws Exception {
         http.authorizeRequests()
                 .antMatchers("/img/**", "/bootstrap/**", "/js/**", "/favicon/**").permitAll()
-                .antMatchers("/panel", "/manager", "/stats").hasRole("ADMIN")
+                .antMatchers("/panel", "/manager", "/stats", "/users", "/users/**").hasRole("ADMIN")
                 .antMatchers("/list", "/add_new_drink").hasAnyRole("ADMIN", "USER", "REGISTERED")
                 .antMatchers("/", "/index", "/signup", "/new_account", "/show-more", "/login/**").permitAll()
-                .antMatchers("/users", "/users/**").permitAll()
                 .anyRequest().authenticated()
                 .and()
                 .formLogin()
